Refactoring: Implemented up to single Product Discount

+++ src/main/java/com/booleanuk/core/Inventory.java
@@ -3,6 +3,7 @@ package com.booleanuk.core;
 import com.booleanuk.core.model.Product;
 import com.booleanuk.core.model.Supplement;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -10,7 +11,7 @@ public class Inventory {
     private Map<String, Product> products;
     private Map<String, Supplement> supplements;
     private static Inventory instance;
-
+    //TODO: consider changing to static (better performance)
     private Inventory() {
         products = new HashMap();
         supplements = new HashMap<>();
@@ -32,9 +33,9 @@ public class Inventory {
 
     public double getPrice(String sku) {
         if (products.get(sku) != null)
-            return products.get(sku).getPrice();
+            return products.get(sku).getPrice().doubleValue();
         if (supplements.get(sku) != null)
-            return supplements.get(sku).getPrice();
+            return supplements.get(sku).getPrice().doubleValue();
         System.out.println("Item not in inventory");
         return -1;
     }
